+++ codbex-payments/events/CustomerPayment/CustomerPaymentToRecordHandler.ts
@@ -1,16 +1,19 @@
-import { PaymentRecordRepository } from "../../gen/codbex-payments/dao/PaymentRecord/PaymentRecordRepository";
+import { PaymentRecordRepository } from "../../gen/codbex-payments/data/PaymentRecord/PaymentRecordRepository";
+import { PaymentRecordEntity } from "../../gen/codbex-payments/data/PaymentRecord/PaymentRecordEntity";
+
+import { EntityEvent } from "@aerokit/sdk/db";
 
 const validate = () => { }
 
-export const trigger = (event) => {
+export const trigger = (event: EntityEvent<PaymentRecordEntity>): void => {
     const PaymentRecordDao = new PaymentRecordRepository();
     const item = event.entity;
     const operation = event.operation;
 
     if (operation === "create") {
         const record = {
-            Date: item.Date,
-            Valor: item.Valor,
+            Date: item.Date!,
+            Valor: item.Valor!,
             CompanyIBAN: item.CompanyIBAN,
             CounterpartyIBAN: item.CounterpartyIBAN,
             CounterpartyName: item.CounterpartyName,
